+++ src/components/relatedDatabank/CharacterCard.module.css
@@ -9,18 +9,19 @@
   box-shadow: 0 0 10px 0 rgba(0, 0, 0, 0.2);
   margin: 20px;
   overflow: hidden;
+  /* position: relative; */
 }
-/* .dent {
-  position: absolute;
-  color: black;
-  display: block;
-  margin-left: 400px;
-  border-top-left-radius: 100em;
-  top: 500px;
-  rotate: 90deg;
-  font-size: 10rem;
-  transform: scaleY(3);
-} */
+
+/*  
+  .dent {
+    border-bottom: 18px solid black;
+    border-left: 18px solid transparent;
+    border-right: 18px solid transparent;
+    width: 150px;
+    position: absolute;
+    top: 663px;
+    right: 70px;
+  } */
 
 .images {
   display: flex;
@@ -42,23 +43,37 @@
   padding: 10px 25px 35px 18px;
 }
 
-/* Small devices (landscape phones, 576px and up) */
-@media (min-width: 680px) {
-}
 
-/* Medium devices (tablets, 768px and up) */
-@media (min-width: 768px) {
-  /* Code here */
-}
+/* XX-Large devices (larger desktops, 850px and up) */
+@media (min-width: 850px) {
+ 
+  .luminous{
+  margin-bottom: 3px
+  }
 
-/* Large devices (desktops, 992px and up) */
-@media (min-width: 992px) {
+/*  
+  .dent {
+    border-bottom: 18px solid black;
+    border-left: 18px solid transparent;
+    border-right: 18px solid transparent;
+    width: 150px;
+    position: absolute;
+    top: 663px;
+    right: 70px;
+  } */
+  .images {
+    width: auto;
+    height: 60vh;
+    /* display: block;
+    float: left; */
+  }
+  .written-info {
+    width: 100%;
+    height: 100%;
+    display: inline;
+    padding: 10px 25px 35px 18px;
+  }
 }
-/* X-Large devices (large desktops, 1200px and up) */
-@media (min-width: 1200px) {
-  /* Code here */
-}
-
 /* XX-Large devices (larger desktops, 1400px and up) */
 @media (min-width: 1125px) {
   .card {
@@ -78,24 +93,19 @@
   margin-bottom: 3px
   }
 
- 
- /* .dent {
+/*  
+  .dent {
+    border-bottom: 18px solid black;
+    border-left: 18px solid transparent;
+    border-right: 18px solid transparent;
+    width: 150px;
     position: absolute;
-
-    color: black;
-    background-color: white;
-    display: block;
-  
-    margin-left: 800px;
-    border-top-left-radius: 100em;
-    top: 400px;
-    rotate: 90deg;
-    font-size: 10rem;
-    transform: scaleY(5);
+    top: 663px;
+    right: 70px;
   } */
   .images {
     width: auto;
-    height: 60%;
+    height: 45vh;
     /* display: block;
     float: left; */
   }
