+++ src/views/IndividualDatabankCard.module.css
@@ -1,0 +1,11 @@
+/* .dent {
+  border-bottom: 15px solid red;
+  border-left: 15px solid transparent;
+  border-right: 15px solid transparent;
+  width: 110px;
+  position: absolute;
+  top: 530px;
+  right: 60px;
+} */
+
+
